+++ vem-comigo-backend/src/entity/caroneiro.entity.ts
@@ -6,6 +6,6 @@ export class Caroneiro extends Usuario {
 	@Column('int', { nullable: false, name: 'qtd_viagens' })
 	qtd_viagens!: number;	
 	
-	@Column('double', { nullable: false, name: 'nota_media' })
+	@Column('float', { nullable: false, name: 'nota_media' })
 	nota_media!: number;
 }
