+++ vem-comigo-backend/src/entity/caronista.entity.ts
@@ -10,7 +10,7 @@ export class Caronista extends Usuario {
 	@Column('int', { nullable: false, name: 'qtd_viagens' })
 	qtd_viagens!: number;	
 	
-	@Column('double', { nullable: false, name: 'nota_media' })
+	@Column('float', { nullable: false, name: 'nota_media' })
 	nota_media!: number;
 
 	@OneToMany(() => Veiculo, veiculo => veiculo.caronista)
